Remove CPC+ as a valid program name and update PcfClinicalDocumentValidator
