--- converter/src/test/java/gov/cms/qpp/conversion/validate/CpcClinicalDocumentValidatorTest.java
@@ -1,240 +1,0 @@
-package gov.cms.qpp.conversion.validate;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-
-import gov.cms.qpp.conversion.Context;
-import gov.cms.qpp.conversion.decode.ClinicalDocumentDecoder;
-import gov.cms.qpp.conversion.model.Node;
-import gov.cms.qpp.conversion.model.TemplateId;
-import gov.cms.qpp.conversion.model.error.Detail;
-import gov.cms.qpp.conversion.model.error.ProblemCode;
-import gov.cms.qpp.conversion.model.error.correspondence.DetailsErrorEquals;
-import gov.cms.qpp.conversion.model.validation.ApmEntityIds;
-
-import java.time.ZonedDateTime;
-import java.util.List;
-
-import static com.google.common.truth.Truth.assertThat;
-import static com.google.common.truth.Truth.assertWithMessage;
-
-class CpcClinicalDocumentValidatorTest {
-
-	private CpcClinicalDocumentValidator cpcValidator;
-	private ApmEntityIds apmEntityIds;
-
-	@BeforeEach
-	void setUp() {
-		apmEntityIds = new ApmEntityIds("test_apm_entity_ids.json", "test_apm_entity_ids.json");
-		cpcValidator = new CpcClinicalDocumentValidator(new Context(apmEntityIds));
-	}
-
-	@AfterEach
-	void cleanUp() {
-		System.clearProperty(CpcClinicalDocumentValidator.END_DATE_VARIABLE);
-		System.clearProperty(CpcClinicalDocumentValidator.CPC_PLUS_CONTACT_EMAIL);
-	}
-
-	@Test
-	void validPracticeSiteAddress() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must have no errors")
-				.that(errors).isEmpty();
-	}
-
-	@Test
-	void missingPracticeSiteAddress() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-		clinicalDocumentNode.removeValue(ClinicalDocumentDecoder.PRACTICE_SITE_ADDR);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must contain error")
-				.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-				.containsExactly(ProblemCode.CPC_PCF_CLINICAL_DOCUMENT_MISSING_PRACTICE_SITE_ADDRESS
-					.format(Context.REPORTING_YEAR));
-	}
-
-	@Test
-	void testCpcPlusMultipleApm() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-
-		// extra APM
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.PRACTICE_ID, "1234567", false);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must validate with the correct error")
-				.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-				.containsExactly(ProblemCode.CPC_PCF_CLINICAL_DOCUMENT_ONLY_ONE_APM_ALLOWED);
-	}
-
-	@Test
-	void testCpcPlusNoApm() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-		clinicalDocumentNode.removeValue(ClinicalDocumentDecoder.PRACTICE_ID);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must validate with the correct error")
-				.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-				.contains(ProblemCode.CPC_PCF_CLINICAL_DOCUMENT_ONLY_ONE_APM_ALLOWED);
-	}
-
-	@Test
-	void testCpcPlusEmptyApm() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.PRACTICE_ID, "");
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-		assertWithMessage("Must validate with the correct error")
-			.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.containsExactly(ProblemCode.CPC_PCF_CLINICAL_DOCUMENT_EMPTY_APM);
-	}
-
-	@Test
-	void testCpcPlusInvalidApm() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.PRACTICE_ID, "PropertyTaxes");
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-		assertWithMessage("Must validate with the correct error")
-			.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.containsExactly(ProblemCode.CPC_PCF_CLINICAL_DOCUMENT_INVALID_APM);
-	}
-
-	@Test
-	void testCpcPlusMissingMeasureSection() {
-		Node clinicalDocumentNode = createCpcPlusClinicalDocument();
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must validate with the correct error")
-				.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-				.containsExactly(ProblemCode.CPC_PCF_CLINICAL_DOCUMENT_ONE_MEASURE_SECTION_REQUIRED);
-	}
-
-	@Test
-	void testCpcPlusSubmissionBeforeEndDate() {
-		System.setProperty(CpcClinicalDocumentValidator.END_DATE_VARIABLE,
-			ZonedDateTime.now(CpcClinicalDocumentValidator.EASTERN_TIME_ZONE).plusYears(3)
-			.format(CpcClinicalDocumentValidator.INPUT_END_DATE_FORMAT));
-		Node clinicalDocument = createValidCpcPlusClinicalDocument();
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocument).getErrors();
-
-		assertThat(errors)
-			.isEmpty();
-	}
-
-	@ParameterizedTest
-	@CsvSource({"meep@mawp.blah, meep@mawp.blah", ", cpcplus@telligen.com"})
-	void testCpcPlusSubmissionAfterEndDate(String systemValue, String expected) {
-		ZonedDateTime endDate = ZonedDateTime.now(CpcClinicalDocumentValidator.EASTERN_TIME_ZONE).minusYears(3);
-		String formattedDate = endDate.format(CpcClinicalDocumentValidator.OUTPUT_END_DATE_FORMAT);
-		if (systemValue != null) {
-			System.setProperty(CpcClinicalDocumentValidator.CPC_PLUS_CONTACT_EMAIL, systemValue);
-		}
-		System.setProperty(CpcClinicalDocumentValidator.END_DATE_VARIABLE, endDate.format(CpcClinicalDocumentValidator.INPUT_END_DATE_FORMAT));
-		Node clinicalDocument = createValidCpcPlusClinicalDocument();
-
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocument).getErrors();
-
-		assertThat(errors)
-			.comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.containsExactly(ProblemCode.CPC_PCF_PLUS_SUBMISSION_ENDED.format(ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME, ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME, formattedDate, ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME, ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME, expected));
-	}
-
-	@Test
-	void testCpcPlusMissingTin() {
-		Node clinicalDocumentNode = createCpcPlusClinicalDocument();
-		clinicalDocumentNode.removeValue(ClinicalDocumentDecoder.TAX_PAYER_IDENTIFICATION_NUMBER);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must validate with the correct error")
-			.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.containsExactly(ProblemCode.CPC_PCF_PLUS_TIN_REQUIRED
-				.format(ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME.toUpperCase()));
-	}
-
-	@Test
-	void testCpcPlusMissingNpi() {
-		Node clinicalDocumentNode = createCpcPlusClinicalDocument();
-		clinicalDocumentNode.removeValue(ClinicalDocumentDecoder.NATIONAL_PROVIDER_IDENTIFIER);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must validate with the correct error")
-			.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.containsExactly(ProblemCode.CPC_PCF_PLUS_NPI_REQUIRED
-				.format(ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME.toUpperCase()));
-	}
-
-	@Test
-	void testCpcPlusMissingCehrt() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-		clinicalDocumentNode.removeValue(ClinicalDocumentDecoder.CEHRT);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must validate with the correct error")
-			.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.containsExactly(ProblemCode.CPC_PCF_MISSING_CEHRT_ID.format(ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME.toUpperCase()));
-	}
-
-	@Test
-	void testWarnWhenContainsIa() {
-		Node clinicalDocumentNode = createCpcPlusClinicalDocument();
-		Node iaSection = new Node(TemplateId.IA_SECTION);
-		clinicalDocumentNode.addChildNode(iaSection);
-		List<Detail> warnings = cpcValidator.validateSingleNode(clinicalDocumentNode).getWarnings();
-
-		assertThat(warnings)
-			.comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.contains(ProblemCode.CPC_PCF_PLUS_NO_IA_OR_PI.format(ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME.toUpperCase()));
-	}
-
-	@Test
-	void testErrorWhenContainsPi() {
-		Node clinicalDocumentNode = createCpcPlusClinicalDocument();
-		Node piSection = new Node(TemplateId.PI_SECTION_V2);
-		clinicalDocumentNode.addChildNode(piSection);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertThat(errors)
-			.comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.contains(ProblemCode.CPC_PLUS_NO_PI);
-	}
-
-	@Test
-	void testCpcPlusDuplicateCehrt() {
-		Node clinicalDocumentNode = createValidCpcPlusClinicalDocument();
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.CEHRT, "XX15EXXXXXXXXXX", false);
-		List<Detail> errors = cpcValidator.validateSingleNode(clinicalDocumentNode).getErrors();
-
-		assertWithMessage("Must validate with the correct error")
-			.that(errors).comparingElementsUsing(DetailsErrorEquals.INSTANCE)
-			.containsExactly(ProblemCode.CPC_PCF_PLUS_DUPLICATE_CEHRT);
-	}
-
-	private Node createValidCpcPlusClinicalDocument() {
-		Node clinicalDocumentNode = createCpcPlusClinicalDocument();
-		addMeasureSectionNode(clinicalDocumentNode);
-		return clinicalDocumentNode;
-	}
-
-	private Node createCpcPlusClinicalDocument() {
-		Node clinicalDocumentNode = new Node(TemplateId.CLINICAL_DOCUMENT);
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.PROGRAM_NAME, ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME);
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.RAW_PROGRAM_NAME, ClinicalDocumentDecoder.CPCPLUS_PROGRAM_NAME);
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.ENTITY_TYPE, "");
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.PRACTICE_SITE_ADDR, "test");
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.PRACTICE_ID, "DogCow");
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.TAX_PAYER_IDENTIFICATION_NUMBER, "123456789");
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.NATIONAL_PROVIDER_IDENTIFIER, "9900000099");
-		clinicalDocumentNode.putValue(ClinicalDocumentDecoder.CEHRT, "XX15EXXXXXXXXXX");
-
-		return clinicalDocumentNode;
-	}
-
-	private void addMeasureSectionNode(Node clinicalDocumentNode) {
-		Node measureSection = new Node(TemplateId.MEASURE_SECTION_V4);
-		clinicalDocumentNode.addChildNode(measureSection);
-	}
-}
