+++ converter/src/test/java/gov/cms/qpp/conversion/validate/PcfClinicalDocumentValidatorTest.java
@@ -30,8 +30,8 @@ public class PcfClinicalDocumentValidatorTest {
 
 	@AfterEach
 	void cleanUp() {
-		System.clearProperty(CpcClinicalDocumentValidator.END_DATE_VARIABLE);
-		System.clearProperty(CpcClinicalDocumentValidator.CPC_PLUS_CONTACT_EMAIL);
+		System.clearProperty(PcfClinicalDocumentValidator.END_DATE_VARIABLE);
+		System.clearProperty(PcfClinicalDocumentValidator.CPC_PLUS_CONTACT_EMAIL);
 	}
 
 	@Test
@@ -104,9 +104,10 @@ public class PcfClinicalDocumentValidatorTest {
 
 	@Test
 	void testCpcPlusSubmissionBeforeEndDate() {
-		System.setProperty(CpcClinicalDocumentValidator.END_DATE_VARIABLE,
-			ZonedDateTime.now(CpcClinicalDocumentValidator.EASTERN_TIME_ZONE).plusYears(3)
-				.format(CpcClinicalDocumentValidator.INPUT_END_DATE_FORMAT));
+		System.setProperty(PcfClinicalDocumentValidator.END_DATE_VARIABLE,
+			ZonedDateTime.now(PcfClinicalDocumentValidator.EASTERN_TIME_ZONE).plusYears(3)
+				.format(PcfClinicalDocumentValidatorconverter/src/test/java/gov/cms/qpp/conversion/validate/PcfClinicalDocumentValidatorTest.java
+					.INPUT_END_DATE_FORMAT));
 		Node clinicalDocument = createPcfClinicalDocumentNodeWithMeasureSection();
 		List<Detail> errors = validator.validateSingleNode(clinicalDocument).getErrors();
